PdfFontCID: Font embedding with used glyphs only for multibyte encodings
This change allows font embedding with used glyphs only
for multibyte encodings. Note of other two bugs being
fixed together with this and disabled debug prints
in PdfFontTTFSubset.cpp.

There is still room for improvement, like with removed
glyphs the output of font widths can be truncated as well,
same as TTF tables (though it's out of my knowledge).


git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/branches/_dev@1540 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5